# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the Padel Backend service.",
        "title": "Padel Backend API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/v1/organizations/{orgID}/courts/{courtID}/reservations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all reservations for a court within a time range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "List reservations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Court ID",
                        "name": "courtID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start time in RFC3339 format",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End time in RFC3339 format",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.ListReservationsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/organizations/{orgID}/courts/{courtID}/reservations/{reservationID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the reservation with the specified ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Get a reservation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Court ID",
                        "name": "courtID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Reservation ID",
                        "name": "reservationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.ReservationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancels the reservation with the specified ID.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Cancel a reservation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reservation ID",
                        "name": "reservationID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cancellation payload",
                        "name": "cancel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.CancelReservationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/organizations/{orgID}/courts/{courtID}/reserve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a reservation for the specified organization and court.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Reserve a court",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "orgID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Court ID",
                        "name": "courtID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reservation payload",
                        "name": "reservation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.ReserveCourtRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_controllers_http.CancelReservationRequest": {
            "type": "object",
            "properties": {
                "cancelledBy": {
                    "description": "CancelledBy is the ID or name of the user performing the cancellation\nexample: \"user_123\"",
                    "type": "string"
                }
            }
        },
...
        "internal_controllers_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Message provides a human-readable description of the error.",
                    "type": "string",
                    "example": "invalid JSON body"
                }
            }
        },
...
        "internal_controllers_http.ListReservationsResponse": {
            "type": "object",
            "properties": {
                "reservations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_controllers_http.ReservationResponse"
                    }
                }
            }
        },
...
        "internal_controllers_http.ReservationResponse": {
            "type": "object",
            "properties": {
                "cancelledBy": {
                    "type": "string",
                    "example": ""
                },
                "courtId": {
                    "type": "string",
                    "example": "court-456"
                },
                "createdAt": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-11-01T10:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "res-123"
                },
                "reservedBy": {
                    "type": "string",
                    "example": "user-789"
                },
                "reservedFrom": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-11-04T18:30Z"
                },
                "reservedTo": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-11-04T19:45Z"
                },
                "status": {
                    "type": "string",
                    "example": "reserved"
                }
            }
        },
        "internal_controllers_http.ReserveCourtRequest": {
            "type": "object",
            "properties": {
                "endTime": {
                    "description": "EndTime is the reservation end timestamp in RFC3339 format\nexample: 2025-11-04T19:45\nformat: date-time",
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-11-04T19:45"
                },
                "startTime": {
                    "description": "StartTime is the reservation start timestamp in RFC3339 format\nexample: 2025-11-04T18:30\nformat: date-time",
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-11-04T18:30"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}